Fixing all testing exceptions and failures, plus pep8 problems. Get it to pass again.
next is to restore 100% coverage
